Split init_genericrc into trustx and trustme specific parts
If we build for trustx, no dnsmasq and mdm connection(stunnel)
is needed in cml. In the trustme case, no tpm2d is needed.
Therefore, this specific services are removed from the
init_generic_cml_debug.rc and put in the seperate
trustme_generic_{me|x}.rc files.
Those files need to be included in the corresponding init files of the
hardware specifc trustme_<device>_cml repositories.